fix(ci): deduplicate CIFAR-10 download scripts and update comprehensive tests workflow
Closes #3742
